Add files via upload
Trick Raiders is a trick-taking game where a player is dealt some cards as a traditional hand and some cards face up on the table which form both their bid and are the only cards which may be used to lead to tricks. Prior to the game commencing, players may add to, leave alone, or remove cards from this face up group.

It is a must-follow trick-taking game with a fixed trump suit.

Players may be eliminated from a hand if it is their turn to lead to a trick and they are out of leading cards or if they need to follow and are out of cards in their hand. Players begin the game with a few life chips and the player with the most remaining when one person is out is the winner.

A few ranks also have special abilities, for instance, when the 8 or 9 of clubs is played and wins a trick, the trump suit changes.

English rules : https://boardgamegeek.com/filepage/249215/trick-raiders-english-rules
BGG page : https://boardgamegeek.com/boardgame/374336/trick-raiders